Move 13 unit test files from tests/ to inline #[cfg(test)] modules in src/
Unit tests for protocol parsing, config defaults, error Display, switch
definitions, serialization, and property-based tests were sitting in
tests/ as separate binaries. Each file produced its own compile+link
cycle, with 13 binaries consuming significant build time on the critical
path.

Moving them into #[cfg(test)] modules in their corresponding src/ files
eliminates these binaries — they now compile as part of the existing
lib(test) target. This reduces test binary count from 55 to 42 and cuts
test compilation CPU by ~42%.

Services affected: ppba-driver (4 files), qhy-focuser (3 files),
phd2-guider (5 files), filemonitor (1 file).
